Tons of fixes for various issues, include `xaudio2_9redist.dll` (RTC will not work without this)

+++ src/main/kotlin/gg/sona/eos/EosPlatform.kt
@@ -246,7 +246,7 @@ public class EosPlatform internal constructor(
         val size = sizePtr.get(ValueLayout.JAVA_INT, 0)
         val buf = arena.allocate(size.toLong())
         sizePtr.set(ValueLayout.JAVA_INT, 0, size)
-        fn.invokeExact(handle, userId.raw, buf, sizePtr)
+        fn.invokeExact(handle, userId.raw, buf, sizePtr) as Int
         buf.getString(0)
     }
